backend/ucx: enable CUDA support
Implement CUDA context tracking in UCX progress thread.
if a VRAM memory is registered - query CUDA context and
set it in the progress thread

Also implement a dynamic way to disable it for testing
purposes. The issue was originally observed on on of
the internal clusters but wasn't reproduced during
development. So ability to disable it would help
to investigate the issue if reproduced.

Signed-off-by: Artem Y. Polyakov <artemp@nvidia.com>